### 2022-05-17 #4399 @editorialbot
[PRE REVIEW]: musclesyneRgies: factorization of electromyographic data in R with sensible defaults
**Submitting author:** @alesantuz (<a href="http://orcid.org/0000-0002-6577-5101">Alessandro Santuz</a>)
**Repository:** https://github.com/alesantuz/musclesyneRgies
**Branch with paper.md** (empty if default branch): master
**Version:** 1.1.3
