### 2022-05-17 #4399 @Kevin-Mattheus-Moerman
@alesantuz I recommend that you have just a plain LICENSE file and have it contain the MIT license. You already have a copyright statement in that so wont need the additional license file one that. I think if you do this then GitHub will also pick up this license properly and render the MIT license "badge" for your repository.
