### 2022-05-18 #4399 @alesantuz
> @alesantuz can you check that potentially missing DOI? ☝️ ? Note you can call `@editorialbot generate pdf` to update the paper and you can call `@editorialbot check references` too.

Thanks @Kevin-Mattheus-Moerman. That book does not appear to have a DOI and the one referenced by editorialbot links to a "book notice" from a journal.
