### 2022-05-18 #4399 @alesantuz
> @alesantuz I recommend that you have just a plain LICENSE file and have it contain the MIT license. You already have a copyright statement in that so wont need the additional license file one that. I think if you do this then GitHub will also pick up this license properly and render the MIT license "badge" for your repository.

@Kevin-Mattheus-Moerman since I am using the repo for the development of the R-package, I am afraid the plain LICENSE file **must** contain only year and copyright holder's name. The trick you mentioned did not work for CRAN submission last February, unfortunately.
