### 2022-05-18 #4399 @Kevin-Mattheus-Moerman
@alesantuz

> Should I also invite someone or shall we wait until we get a reply? Thank you!

No only the AEiC can invite editors, and only the handling editor can invite reviewers. However, you may, at this point if you wish, help suggest reviewers. When you do so please do not use the @ in the github handles of suggested reviewers, so they are not tagged yet.
