### 2022-05-19 #4399 @fabian-s
@alesantuz

thanks for your suggestions, that is always very helpful. let's see how the first batch responds...

in the meantime, you could adress this:
```
MISSING DOIs

- 10.1016/0006-8993(69)90278-9 may be a valid DOI for title: The co-ordination and regulation of movements
```
