### 2022-05-19 7ebf94f
Updated README and `Hurst` help
