### 2022-05-19 5319d40
Updated paper and bibliography
