+++ paper/paper.bib
@@ -187,3 +187,103 @@
    volume = {2018},
    year = {2018},
 }
+
+@article{Martino2014,
+   author = {Giovanni Martino and Yury P. Ivanenko and M. Serrao and A. Ranavolo and Andrea D'Avella and F. Draicchio and C. Conte and C. Casali and Francesco Lacquaniti},
+   doi = {10.1152/jn.00275.2014},
+   isbn = {1522-1598 (Electronic)\r0022-3077 (Linking)},
+   issn = {0022-3077},
+   issue = {11},
+   journal = {Journal of Neurophysiology},
+   month = {12},
+   pages = {2810-2821},
+   pmid = {25185815},
+   title = {Locomotor patterns in cerebellar ataxia},
+   volume = {112},
+   url = {http://www.ncbi.nlm.nih.gov/pubmed/25185815 http://jn.physiology.org/cgi/doi/10.1152/jn.00275.2014},
+   year = {2014},
+}
+
+@article{Higuchi1988,
+   author = {Tomoyuki Higuchi},
+   doi = {10.1016/0167-2789(88)90081-4},
+   issn = {01672789},
+   issue = {2},
+   journal = {Physica D: Nonlinear Phenomena},
+   month = {6},
+   pages = {277-283},
+   title = {Approach to an irregular time series on the basis of the fractal theory},
+   volume = {31},
+   url = {https://linkinghub.elsevier.com/retrieve/pii/0167278988900814},
+   year = {1988},
+}
+
+@article{Santuz2020a,
+   author = {Alessandro Santuz and Turgay Akay},
+   doi = {10.1152/jn.00360.2020},
+   issn = {0022-3077},
+   issue = {4},
+   journal = {Journal of Neurophysiology},
+   month = {10},
+   pages = {1083-1091},
+   title = {Fractal analysis of muscle activity patterns during locomotion: pitfalls and how to avoid them},
+   volume = {124},
+   url = {https://journals.physiology.org/doi/10.1152/jn.00360.2020},
+   year = {2020},
+}
+
+@article{Hurst1951,
+   author = {Harold Edwin Hurst},
+   doi = {10.1061/TACEAT.0006518},
+   journal = {Transactions of the American Society of Civil Engineers},
+   pages = {770-808},
+   title = {Long-term storage capacity of reservoirs},
+   volume = {116},
+   year = {1951},
+}
+
+@article{Rosenstein1993,
+   author = {Michael T. Rosenstein and James J. Collins and Carlo J. De Luca},
+   doi = {10.1016/0167-2789(93)90009-P},
+   isbn = {0167-2789(92)00033-6},
+   issn = {01672789},
+   issue = {1-2},
+   journal = {Physica D},
+   month = {5},
+   pages = {117-134},
+   pmid = {108},
+   title = {A practical method for calculating largest Lyapunov exponents from small data sets},
+   volume = {65},
+   url = {http://linkinghub.elsevier.com/retrieve/pii/016727899390009P},
+   year = {1993},
+}
+
+@article{Santuz2020b,
+   author = {Alessandro Santuz and Leon Brüll and Antonis Ekizos and Arno Schroll and Nils Eckardt and Armin Kibele and Michael Schwenk and Adamantios Arampatzis},
+   doi = {10.1016/j.isci.2019.100796},
+   issn = {25890042},
+   issue = {1},
+   journal = {iScience},
+   month = {1},
+   pages = {100796},
+   title = {Neuromotor Dynamics of Human Locomotion in Challenging Settings},
+   volume = {23},
+   url = {https://www.cell.com/iscience/fulltext/S2589-0042(19)30542-5},
+   year = {2020},
+}
+
+@article{Kang2006,
[…]
