### 2022-05-19 #4399 @alesantuz
As mentioned above, the suggested DOI to replace the missing one does not refer to the correct publication and links instead to the following, which is just a "Book notice" from the journal Brain Research:

> The co-ordination and regulation of movements. (1969). Brain Research, 13(3), 624. doi:10.1016/0006-8993(69)90278-9

I still could not find a valid DOI for Bernstein's book.
