### 2022-05-19 #4399 @fabian-s
:wave: @modenaxe @davidruegamer @cpizzolato

Would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
