### 2022-05-23 #4399 @modenaxe
Hi @fabian-s this submission looks great but I am afraid I don't have the time for reviewing it plus my R is few years rusty. I will try to think about alternative reviewers, although my main suggestion would have been @cpizzolato!
