### 2022-05-23 168f1d9
`rawdata` can now handle RData files saved with `save()`
