+++ man/Hurst.Rd
@@ -24,7 +24,7 @@ Hurst exponent
 \details{
 \code{Hurst} calculates the Hurst exponent based on the R/S approach as in Hurst (1951).
 For motor primitives, the minimum window length should be bigger than the period
-(i.e., the length of each cycle), as reported in Santuz, Akay (2020).
+(i.e. the length of each cycle), as reported in Santuz, Akay (2020).
 }
 \examples{
 # Measurements of the annual flow of the river Nile at Aswan
