### 2022-05-24 af7811c
Update subsetEMG.R
read cycles as data frame
