### 2022-05-24 672a35b
To deal with single-column cycles
