+++ NEWS.md
@@ -1,3 +1,15 @@
+# musclesyneRgies 1.1.3.9004 (development version)
+### How to install
+```
+install.packages("devtools")
+devtools::install_github("alesantuz/musclesyneRgies")
+```
+### How to use
+README and vignettes are available both on [CRAN](https://CRAN.R-project.org/package=musclesyneRgies) and on [GitHub](https://github.com/alesantuz/musclesyneRgies).
+
+### What's changed
+- Updated tests to check for compatibility with single-column cycles.
+
 # musclesyneRgies 1.1.3.9003 (development version)
 ### How to install
 ```
