### 2022-05-25 a3448f6
Updated NEWS
