+++ NEWS.md
@@ -1,8 +1,8 @@
 # musclesyneRgies 1.1.3.9004 (development version)
 ### How to install
 ```
-install.packages("devtools")
-devtools::install_github("alesantuz/musclesyneRgies")
+install.packages("remotes")
+remotes::install_github("alesantuz/musclesyneRgies")
 ```
 ### How to use
 README and vignettes are available both on [CRAN](https://CRAN.R-project.org/package=musclesyneRgies) and on [GitHub](https://github.com/alesantuz/musclesyneRgies).
