### 2022-05-29 #4439 @editorialbot
[REVIEW]: musclesyneRgies: factorization of electromyographic data in R with sensible defaults
**Submitting author:** @alesantuz (<a href="http://orcid.org/0000-0002-6577-5101">Alessandro Santuz</a>)
**Repository:** https://github.com/alesantuz/musclesyneRgies
**Branch with paper.md** (empty if default branch): master
**Version:** v1.2.3
**Archive:** 10.5281/zenodo.6767313.
