### 2022-05-29 #4439 @fabian-s
@alesantuz, the paper (~1600 words) is currently rather long for JOSS, we aim for <1000 words.

You could consider moving the "typical workflow" section into a vignette for your package and replacing that section in your article with a more high-level description of the functionality, for example.
