### 2022-05-29 #4439 @alesantuz
> @alesantuz, the paper (~1600 words) is currently rather long for JOSS, we aim for <1000 words.
>
> You could consider moving the "typical workflow" section into a vignette for your package and replacing that section in your article with a more high-level description of the functionality, for example.
>
> EDIT: should have been more clear: "<1000 words" is not a hard, mandatory thing, but we would prefer if you could shorten your paper by some amount.

@fabian-s I can certainly do that, since "workflow" is already a vignette and is more detailed than the text in the paper. I will cut that section down to the bone, perhaps leaving only the bullet list and the compact code example (lines 47 to 57). I'd leave the figure though, if that's ok.
