### 2022-05-29 #4439 @fabian-s
> "workflow" is already a vignette which is anyway more detailed than the text in the paper.

cool, so that won't be much trouble anyways.

> I will cut that section down to the bone, perhaps leaving only the bullet list and the compact code example (lines 47 to 57). I'd leave the figure though, if that's ok.

perfect!
