### 2022-05-29 f09e316
Reduced "Typical workflow" section of paper
