+++ paper/paper.md
@@ -36,12 +36,14 @@ The great amount of muscles and joints in the body of vertebrate animals makes t
 
 The typical workflow when using `musclesyneRgies` consists of six main steps:
 
-1. Data preparation
-2. Raw data processing
-3. Synergy extraction
-4. Synergy classification
+1. Data preparation (to read raw data sets and covert them into the needed format)
+2. Raw data processing (e.g. rectification, filtering, time-normalisation, etc.)
+3. Synergy extraction (via NMF)
+4. Synergy classification (via k-means)
 5. Synergy analysis
-6. Plots (available at each of the previous steps).
+    i. Linear methods: full width at half maximum and centre of activity [@Martino2014]
+    ii. Non-linear methods: local complexity or Higuchi's fractal dimension [@Higuchi1988; @Santuz2020a], global complexity or Hurst exponent [@Hurst1951; @Santuz2020a], short-term maximum Lyapunov exponents [@Rosenstein1993; @Santuz2020b; @Kang2006]
+6. Plots (available at each of the previous steps, see \autoref{fig:syns} for an example).
 
 Using the native pipe operator (R >= `4.1.0` is required), a typical analysis pipeline can be synthetically written as follows:
 
@@ -52,137 +54,8 @@ SYNS_classified <- lapply(RAW_DATA, filtEMG) |>       # Filter raw data
   classify_kmeans()                                   # Synergy classification
 ```
 
-Extensive documentation is available on [GitHub](https://github.com/alesantuz/musclesyneRgies) and the [Comprehensive `R` Archive Network](https://CRAN.R-project.org/package=musclesyneRgies). Below, a typical analysis workflow is provided as an example.
+Defaults are specifically targeted at the analysis of human and mouse locomotion, but they can be flexibly overridden by specifying the arguments of the relevant functions. Extensive documentation is available on [GitHub](https://github.com/alesantuz/musclesyneRgies) and the [Comprehensive `R` Archive Network](https://CRAN.R-project.org/package=musclesyneRgies).
 
-## Data preparation
-
-The data set must be a list of objects of class `EMG`, each of which is a list with two elements:
-
-- `cycles` data frame containing cycle timings, with as many columns as many cycle subdivisions are wanted (column names are not needed but can be present)
-- `emg` data frame containing raw EMG data in columns, with the first column being time information in the same units as in `cycles` (column names must be present and reflect the muscle names).
-
-Those two elements should look similar to the following:
-
-```r
-library(musclesyneRgies)
-data("RAW_DATA")
-head(RAW_DATA[[1]]$cycles)
-```
-
-```
-##      V1    V2
-## 1 1.414 2.074
-## 2 2.448 3.115
-## 3 3.488 4.141
-## 4 4.515 5.168
-## 5 5.549 6.216
-## 6 6.596 7.249
-```
-
-```r
-head(RAW_DATA[[1]]$emg[, 1:6])
-```
-
-```
-##    time         ME        MA       FL        RF        VM
-## 1 0.014   0.201416 -6.445313 22.65930 -0.100708 -0.906372
-## 2 0.015  -2.316284 -0.100708 24.16992  1.812744 -1.913452
-## 3 0.016  -7.351685 -7.150269 23.46497  0.704956 -5.337524
-## 4 0.017  -5.538940 -3.222656 27.49329  5.236816 -4.330444
-## 5 0.018 -10.675049 -5.740356 23.16284 -0.704956  2.014160
-## 6 0.019 -12.487793 -3.927612 19.94019  2.014160 -5.136108
-```
-
-It is also possible to read directly from ASCII files such as tab-separated txt or comma-separated csv and then, using the function `rawdata`, automatically create the list of objects of class `EMG` needed for the subsequent steps.
-
-## Raw data processing
-
[…]
