+++ paper/paper_example.md
@@ -1,228 +1,0 @@
----
-  title: 'Gala: A Python package for galactic dynamics'
-tags:
-  - Python
-- astronomy
-- dynamics
-- galactic dynamics
-- milky way
-authors:
-  - name: Adrian M. Price-Whelan^[co-first author] # note this makes a footnote saying 'co-first author'
-orcid: 0000-0002-6577-5101
-affiliation: "1, 2, 3"
-affiliations:
-  - name: Department of Training and Movement Sciences, Humboldt-Universität zu Berlin, Berlin, Germany
-index: 1
-- name: Berlin School of Movement Science, Humboldt-Universität zu Berlin, Berlin, Germany
-index: 2
-- name: Institute for Biomechanics, ETH Zurich, Zurich, Switzerland
-index: 3
-date: 10 February 2022
-bibliography: paper.bib
-
----
-
-  # Summary
-
-  A
-
-# Statement of need
-
-A
-
-# Typical workflow
-
-A
-
-## Data preparation
-
-A
-
-## Raw data processing
-
-A
-
-## Synergy extraction
-
-A
-
-## Synergy classification
-
-A
-
-## Plots
-
-A
-
-
-
-
-# Availability
-
-A stable release of the `musclesyneRgies` R package is freely available via the
-[Comprehensive `R` Archive Network](https://CRAN.R-project.org/package=musclesyneRgies).
-Its development version can be found on
-[GitHub](https://github.com/alesantuz/musclesyneRgies). Documentation and examples
-are contained in each version's manual pages, vignette.
-
-
-# Acknowledgments
-
-grateful to ... and we would like to thank reviewers and the editor for taking the time and effort necessary to review the manuscript.
-
-# References
-
-A
-
-
-
-
-
-
-# Summary
-
-A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.
-
-EXAMPLE:
-The forces on stars, galaxies, and dark matter under external gravitational
-fields lead to the dynamical evolution of structures in the universe. The orbits
-of these bodies are therefore key to understanding the formation, history, and
-future state of galaxies. The field of "galactic dynamics," which aims to model
-the gravitating components of galaxies to study their structure and evolution,
-is now well-established, commonly taught, and frequently used in astronomy.
-Aside from toy problems and demonstrations, the majority of problems require
-efficient numerical tools, many of which require the same base code (e.g., for
-performing numerical orbit integration).
-
-# Statement of need
-
-A Statement of Need section that clearly illustrates the research purpose of the software.
-
-EXAMPLE:
-`Gala` is an Astropy-affiliated Python package for galactic dynamics. Python
-enables wrapping low-level languages (e.g., C) for speed without losing
-flexibility or ease-of-use in the user-interface. The API for `Gala` was
-designed to provide a class-based and user-friendly interface to fast (C or
-Cython-optimized) implementations of common operations such as gravitational
-potential and force evaluation, orbit integration, dynamical transformations,
-and chaos indicators for nonlinear dynamics. `Gala` also relies heavily on and
-interfaces well with the implementations of physical units and astronomical
-coordinate systems in the `Astropy` package [@astropy] (`astropy.units` and
-`astropy.coordinates`).
-
-`Gala` was designed to be used by both astronomical researchers and by
-students in courses on gravitational dynamics or astronomy. It has already been
[…]

[…]
