### 2022-06-15 04a216c
Added test data set link and example to README
