### 2022-06-15 d1452ce
Added definition of muscle synergies to README
