### 2022-06-15 94e8287
Added definition of modules and primitives to README
