### 2022-06-15 47f0d1f
Improved documentation of `classify_kmeans`
