### 2022-06-15 ba37737
Improved documentation of `HFD`, `Hurst`
