### 2022-06-15 f1039ed
Added test for `classify_kmeans`
