### 2022-06-16 alesantuz/musclesyneRgies#69 @fabian-s
JOSS review: code quality
automated code quality checks from `goodpractice::gp` indicate some potential improvements, specifically to `classify_kmeans`:

- improving code coverage of unit tests (use `covr` to see which parts of the code are currently not covered)
- use TRUE/FALSE not T/F
- Use `seq_len(..)` or `seq_along(..)`  instead of `1:length(..)`

```
> goodpractice::gp("musclesyneRgies")
[....]

It is good practice to

  ✖ write unit tests for all functions, and all package code in
    general. 72% of code lines are covered by test cases.

    R/classify_kmeans.R:56:NA
    R/classify_kmeans.R:57:NA
    R/classify_kmeans.R:58:NA
    R/classify_kmeans.R:59:NA
    R/classify_kmeans.R:67:NA
    ... and 340 more lines

  ✖ write short and simple functions. These functions have high
    cyclomatic complexity:classify_kmeans (103).

    R/classify_kmeans.R:4:1
    R/classify_kmeans.R:5:1
    R/classify_kmeans.R:6:1
    R/classify_kmeans.R:9:1
    R/classify_kmeans.R:10:1
    ... and 173 more lines

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...)
    and 1:NCOL(...) expressions. They are error prone and result 1:0 if
    the expression on the right hand side is zero. Use seq_len() or
    seq_along() instead.

    R/classify_kmeans.R:226:16
    R/classify_kmeans.R:318:37
    R/classify_kmeans.R:319:37
    R/classify_kmeans.R:353:15
    R/classify_kmeans.R:404:48
    ... and 10 more lines

  ✖ avoid 'T' and 'F', as they are just variables which are set
    to the logicals 'TRUE' and 'FALSE' by default, but are not reserved
    words and hence can be overwritten by the user.  Hence, one should
    always use 'TRUE' and 'FALSE' for the logicals.

    R/classify_kmeans.R:NA:NA
    R/classify_kmeans.R:NA:NA
    R/classify_kmeans.R:NA:NA
    R/classify_kmeans.R:NA:NA
    R/classify_kmeans.R:NA:NA
    ... and 5 more lines
```

These are just friendly suggestions, not mandatory for JOSS acceptance.
