### 2022-06-16 #4439 @fabian-s
@vbaliga thanks a *lot* for your careful review and constructive comments - if you feel the issues you pointed out have been addressed satisfactorily, please check off the respective items in your checklist.
Re the LICENSE/LICENSE.md issue: as @alesantuz explained, this is standard operating procedure for CRAN packages.

@alesantuz I have some additional suggestions that should hopefully be easy to deal with if you choose to do so, see repo.
