### 2022-06-16 7d4c4b8
Cleaned code
