+++ R/classify_kmeans.R
@@ -136,7 +136,7 @@ classify_kmeans <- function(x,
     # Set zeroes to smallest non-negative entry
     temp <- y
     temp[temp == 0] <- Inf
-    y[y == 0] <- min(temp, na.rm = T)
+    y[y == 0] <- min(temp, na.rm = TRUE)
     # Normalise to maximum
     y <- y / max(y)
 
@@ -170,7 +170,7 @@ classify_kmeans <- function(x,
       break
     }
     withinss_interp <- data.frame(
-      xx = c(1:(muscle_num - iter + 1)),
+      xx = 1:(muscle_num - iter + 1),
       yy = withinss[iter:(muscle_num)]
     )
 
@@ -221,12 +221,12 @@ classify_kmeans <- function(x,
   )
 
   # Calculate mutual score squared residuals and find minimum
-  perms <- gtools::permutations(nrow(geoms_P), r = 2, repeats.allowed = T)
+  perms <- gtools::permutations(nrow(geoms_P), r = 2, repeats.allowed = TRUE)
   resids <- numeric()
-  for (perm in 1:nrow(perms)) {
+  for (perm in seq_len(nrow(perms))) {
     resids[perm] <- (geoms_P$score[perms[perm, 1]] - geoms_M$score[perms[perm, 2]])^2
   }
-  perms <- perms[sort(resids, decreasing = F, index.return = T)$ix, ]
+  perms <- perms[sort(resids, decreasing = FALSE, index.return = TRUE)$ix, ]
   perms <- data.frame(perms[-which(duplicated(perms[, 1])), ])
   colnames(perms) <- c("old", "new")
 
@@ -308,15 +308,15 @@ classify_kmeans <- function(x,
 
   # Create ordering rule
   order_rule <- data.frame(
-    old = c(1:clust_num),
+    old = 1:clust_num,
     new = order(apply(mean_P, 1, CoA))
   )
 
   # Apply new order to mean curves
   mean_P <- mean_P[order_rule$new, ]
   mean_M <- mean_M[order_rule$new, ]
-  rownames(mean_P) <- paste0("Syn", 1:nrow(mean_P))
-  rownames(mean_M) <- paste0("Syn", 1:nrow(mean_M))
+  rownames(mean_P) <- paste0("Syn", seq_len(nrow(mean_P)))
+  rownames(mean_M) <- paste0("Syn", seq_len(nrow(mean_M)))
   colnames(mean_M) <- colnames(temp_M)
 
   # Apply new order to all
@@ -350,7 +350,7 @@ classify_kmeans <- function(x,
       graphics::barplot(mean_M[syn, ])
       # Plot motor primitives
       plot(
-        x = c(1:ncol(mean_P)), y = mean_P[syn, ],
+        x = seq_len(ncol(mean_P)), y = mean_P[syn, ],
         ty = "l", main = paste0("Synergy ", syn),
         xlab = "", ylab = "",
         xaxt = "n", yaxt = "n", lwd = 2
@@ -390,19 +390,19 @@ classify_kmeans <- function(x,
           orders_new[cc] <- pp
         }
         orders_new <- as.numeric(orders_new)
-        orders_new <- sort.int(orders_new, index.return = T)$ix
+        orders_new <- sort.int(orders_new, index.return = TRUE)$ix
 
         # Re-create ordering rule
         order_rule <- data.frame(
-          old = c(1:clust_num),
+          old = 1:clust_num,
           new = orders_new
         )
 
         # Make new plots for checking
         mean_P_temp <- mean_P[order_rule$new, ]
         mean_M_temp <- mean_M[order_rule$new, ]
-        rownames(mean_P_temp) <- paste0("Syn", 1:nrow(mean_P_temp))
[…]
