### 2022-06-16 b7bfcde
- Improved tests for `classify_kmeans`
