+++ tests/testthat/test-classify_kmeans.R
@@ -2,7 +2,7 @@ test_that("k-means classification is robust", {
   data("SYNS")
 
   for (tt in 1:5) {
-    SYNS_classified <- classify_kmeans(SYNS)
+    SYNS_classified <- musclesyneRgies::classify_kmeans(SYNS)
     ranks <- unique(
       unlist(
         lapply(
@@ -22,13 +22,13 @@ test_that("k-means classification is robust", {
 rnd_trial <- sample(seq_len(length(SYNS)), size = 1)
 SYNS_err <- SYNS
 SYNS_err[[rnd_trial]]$P <- SYNS_err[[rnd_trial]]$P[-which(SYNS_err[[rnd_trial]]$P$time > 150), ]
-expect_error(classify_kmeans(SYNS_err))
+expect_error(musclesyneRgies::classify_kmeans(SYNS_err))
 
 # Check for same amount of muscles
 rnd_trial <- sample(seq_len(length(SYNS)), size = 1)
 SYNS_err <- SYNS
 SYNS_err[[rnd_trial]]$P$MA <- NULL
 SYNS_err[[rnd_trial]]$M <- SYNS_err[[rnd_trial]]$M[-grep("MA", rownames(SYNS_err[[rnd_trial]]$M)), ]
-expect_error(classify_kmeans(SYNS_err))
+expect_error(musclesyneRgies::classify_kmeans(SYNS_err))
 
 })
