+++ tests/testthat/test-HFD.R
@@ -6,4 +6,9 @@ test_that("HFD works", {
     expected = 1.05,
     tolerance = 0.02
   )
+
+  expect_error(musclesyneRgies::HFD(primitive, k = "a"))
+
+  primitive$signal[5] <- "A"
+  expect_error(musclesyneRgies::HFD(primitive))
 })
