### 2022-06-16 b629f31
Added more plot tests
