### 2022-06-16 alesantuz/musclesyneRgies#69 @alesantuz
Thanks, I'll keep the 100-character width, all the rest is going to be addressed soon.
