### 2022-06-17 #4439 @vbaliga
Hi @alesantuz and @fabian-s

I'm quite happy with the edits that Ale made regarding my comments, and have checked off the corresponding items in the checklist. It's great to see this type of work being published open-source and I think it will be a fine addition to JOSS. Please feel free to let me know if you need anything else from me

Best regards,
Vikram 🐢
