### 2022-06-20 6416296
Increment version number to 1.2.0
