### 2022-06-20 793fc89
Version ready for JOSS paper
