### 2022-06-20 #4439 @alesantuz
> Not really, no, I may have misunderstood your " dive into your great suggestions" and "all the rest is going to be addressed soon" remarks which is why I did not move forward with this yet. I thought your remarks meant that you were about to adress all of the points in [alesantuz/musclesyneRgies#69](https://github.com/alesantuz/musclesyneRgies/issues/69) and wanted to wait until you had done so, but now I understand you meant this more as "I'll add some unit tests, and eventually maybe deal with the rest".

I added the most relevant unit tests while leaving open some that in my opinion are not quite worth of testing (e.g. I suppose that all those `if(interactive())` don't actually need unit testing and are there only because of the CRAN tests, please correct me if I am wrong). All the rest has been fixed in version [1.1.3.9006](https://github.com/alesantuz/musclesyneRgies/releases/tag/v1.1.3.9006) (e.g. avoid using 1:ncol() or similar, write TRUE or FALSE in full, etc.).

> * [x]  Make a tagged release of your software, and list the version tag of the archived version here.
> * [x]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
> * [x]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
> * [x]  Please list the DOI of the archived version here.

Version is archived on GitHub [here](https://github.com/alesantuz/musclesyneRgies/releases/tag/v1.2.0) and on Zenodo [here](https://doi.org/10.5281/zenodo.6670566) with DOI 10.5281/zenodo.6670567.
