### 2022-06-20 #4439 @fabian-s
@alesantuz
spell check was mostly fine too, maybe consider using "center" instead of "centre" in l. 44, since you use american spelling everywhere else?

Also please add that missing DOI editorialbot suggested, seems to be the right one for that old reference:
![image](https://user-images.githubusercontent.com/998541/174617353-8cf3f319-38d0-41ef-9553-61ba37f897b0.png)
