+++ paper/paper.Rmd
@@ -36,15 +36,19 @@ The great amount of muscles and joints in the body of vertebrate animals makes t
 
 # Typical workflow
 
-The typical workflow when using `musclesyneRgies` consists of five main steps:
+The typical workflow when using `musclesyneRgies` consists of six main steps:
 
-1. Data preparation
-2. Raw data processing
-3. Synergy extraction
-4. Synergy classification
-5. Plots (available at each of the previous steps).
+1. Data preparation (to read raw data sets and covert them into the needed format)
+2. Raw data processing (e.g. rectification, filtering, time-normalization, etc.)
+3. Synergy extraction (via NMF)
+4. Synergy classification (via k-means)
+5. Synergy analysis
+    i. Linear methods: full width at half maximum and center of activity [@Martino2014]
+    ii. Non-linear methods: local complexity or Higuchi's fractal dimension [@Higuchi1988; @Santuz2020a], global complexity or Hurst exponent [@Hurst1951; @Santuz2020a], short-term maximum Lyapunov exponents [@Rosenstein1993; @Santuz2020b; @Kang2006]
+6. Plots (available at each of the previous steps, see \autoref{fig:syns} for an example).
 
 Using the native pipe operator (R >= `4.1.0` is required), a typical analysis pipeline can be synthetically written as follows:
+
 ```{r eval = FALSE}
 SYNS_classified <- lapply(RAW_DATA, filtEMG) |>       # Filter raw data
   lapply(function(x) normEMG(x, cycle_div = 100)) |>  # Time-normalization to 100 points
@@ -52,111 +56,16 @@ SYNS_classified <- lapply(RAW_DATA, filtEMG) |>       # Filter raw data
   classify_kmeans()                                   # Synergy classification
 ```
 
-Extensive documentation is available on [GitHub](https://github.com/alesantuz/musclesyneRgies) and the [Comprehensive `R` Archive Network](https://CRAN.R-project.org/package=musclesyneRgies). Below, a typical analysis workflow is provided as an example.
+Defaults are specifically targeted at the analysis of human and mouse locomotion, but they can be flexibly overridden by specifying the arguments of the relevant functions. Extensive documentation is available on [GitHub](https://github.com/alesantuz/musclesyneRgies) and the [Comprehensive `R` Archive Network](https://CRAN.R-project.org/package=musclesyneRgies).
 
-## Data preparation
-
-The data set must be a list of objects of class `EMG`, each of which is a list with two elements:
-
-- `cycles` data frame containing cycle timings, with as many columns as many cycle subdivisions are wanted (column names are not needed but can be present)
-- `emg` data frame containing raw EMG data in columns, with the first column being time information in the same units as in `cycles` (column names must be present and reflect the muscle names).
-
-Those two elements should look similar to the following:
-```{r}
-library(musclesyneRgies)
-data("RAW_DATA")
-head(RAW_DATA[[1]]$cycles)
-head(RAW_DATA[[1]]$emg[, 1:6])
-```
-
-It is also possible to read directly from ASCII files such as tab-separated txt or comma-separated csv and then, using the function `rawdata`, automatically create the list of objects of class `EMG` needed for the subsequent steps.
-
-## Raw data processing
-
-Raw EMG is commonly rectified and filtered before factorization. If the length of the data set needs reduction, the function `subsetEMG` can help to remove unnecessary data points. If no subsetting is needed, raw data can be filtered as follows:
-
-```{r eval = FALSE}
-filtered_EMG <- lapply(
-  RAW_DATA,
-  function(x) {
-    filtEMG(x)
-  }
-)
-```
-
[…]
