### 2022-06-20 #4439 @alesantuz
> @alesantuz spell check was mostly fine too, maybe consider using "center" instead of "centre" in l. 44, since you use american spelling everywhere else?

Thank you, I also corrected a couple of other instances where there was a mixup between Brit and American English. Should I update Zenodo and/or create a new version after the spell check?

About the DOI and as mentioned in the pre-review:
> the suggested DOI to replace the missing one does not refer to the correct publication and links instead to the following, which is just a ["Book notice" from the journal Brain Research](https://www.sciencedirect.com/science/article/pii/0006899369902789)

I still could not find a DOI that actually fits this book.
