+++ NEWS.md
@@ -1,3 +1,5 @@
+# musclesyneRgies 1.2.1
+
 # musclesyneRgies 1.2.0
 ### How to install
 ```
