### 2022-06-20 #4439 @alesantuz
No worries @fabian-s. Version 1.2.1 is archived on GitHub [here](https://github.com/alesantuz/musclesyneRgies/releases/tag/v1.2.1) and on Zenodo [here](https://doi.org/10.5281/zenodo.6670566) with DOI 10.5281/zenodo.6671216.
