### 2022-06-20 #4439 @fabian-s
note to EiCs:
"missing" DOI suggested above is wrong, the paper is from 1967 and does not have a DOI.
