### 2022-06-21 04a93fc
Corrected JOSS paper bibliography
