+++ paper/paper.bib
@@ -3,8 +3,8 @@
     author = {{R Core Team}},
     organization = {R Foundation for Statistical Computing},
     address = {Vienna, Austria},
-    year = {2022},
     url = {https://www.R-project.org/},
+    year = {2022}
 }
 
 @book{Bernstein1967,
@@ -15,7 +15,7 @@
    pages = {196},
    publisher = {Pergamon Press Ltd.},
    title = {The co-ordination and regulation of movements},
-   year = {1967},
+   year = {1967}
 }
 
 @article{Tresch2009,
@@ -31,7 +31,7 @@
    title = {The case for and against muscle synergies},
    volume = {19},
    url = {http://www.pubmedcentral.nih.gov/articlerender.fcgi?artid=2818278&tool=pmcentrez&rendertype=abstract},
-   year = {2009},
+   year = {2009}
 }
 
 @article{Cheung2021,
@@ -48,7 +48,7 @@
    title = {Approaches to revealing the neural basis of muscle synergies: a review and a critique},
    volume = {125},
    url = {https://journals.physiology.org/doi/10.1152/jn.00625.2019},
-   year = {2021},
+   year = {2021}
 }
 
 @article{Bizzi2013,
@@ -62,7 +62,7 @@
    title = {The neural origin of muscle synergies},
    volume = {7},
    url = {http://www.pubmedcentral.nih.gov/articlerender.fcgi?artid=3638124&tool=pmcentrez&rendertype=abstract http://journal.frontiersin.org/article/10.3389/fncom.2013.00051/abstract},
-   year = {2013},
+   year = {2013}
 }
 
 @article{Lee1999,
@@ -78,7 +78,7 @@
    title = {Learning the parts of objects by non-negative matrix factorization},
    volume = {401},
    url = {http://www.ncbi.nlm.nih.gov/pubmed/10548103},
-   year = {1999},
+   year = {1999}
 }
 
 @article{Tresch1999,
@@ -92,7 +92,7 @@
    pmid = {10195201},
    title = {The construction of movement by the spinal cord},
    volume = {2},
-   year = {1999},
+   year = {1999}
 }
 
 @article{Bruton2018,
@@ -107,7 +107,7 @@
    title = {Synergies in coordination: a comprehensive overview of neural, computational, and behavioral approaches},
    volume = {120},
    url = {https://www.physiology.org/doi/10.1152/jn.00052.2018},
-   year = {2018},
+   year = {2018}
 }
 
 @article{Rabbi2020,
@@ -116,7 +116,7 @@
    journal = {Scientific Reports},
    pages = {1-11},
    title = {Non-negative matrix factorisation is the most appropriate method for extraction of muscle synergies in walking and running},
-   year = {2020},
+   year = {2020}
 }
 
 @article{Devarajan2014,
@@ -131,7 +131,7 @@
    title = {On nonnegative matrix factorization algorithms for signal-dependent noise with application to electromyography data},
    volume = {26},
    url = {http://www.ncbi.nlm.nih.gov/pubmed/24684448},
-   year = {2014},
+   year = {2014}
 }
 
 @article{Ebied2018,
@@ -145,7 +145,7 @@
    title = {Evaluation of matrix factorisation approaches for muscle synergy extraction},
    volume = {57},
    url = {https://doi.org/10.1016/j.medengphy.2018.04.003},
-   year = {2018},
+   year = {2018}
 }
 
 @article{Oliveira2014,
@@ -161,7 +161,7 @@
    title = {Motor modules of human locomotion: influence of EMG averaging, concatenation, and number of step cycles},
    volume = {8},
[…]
