### 2022-06-21 #4439 @alesantuz
@fabian-s I tried to understand where that error comes from and there might have been too many commas in the .bib file. I made a new version, 1.2.2, which is archived on GitHub [here](https://github.com/alesantuz/musclesyneRgies/releases/tag/v1.2.2) and on Zenodo [here](https://doi.org/10.5281/zenodo.6670566) with DOI 10.5281/zenodo.6673110.
