### 2022-06-21 #4439 @fabian-s
@openjournals/joss-eics :

- "missing" DOI suggested above is wrong, the paper is from 1967 and does not have a DOI.
- for some reason this cannot be processed by editorialbot, we don't know what else to try.... could you look into this?
