### 2022-06-27 #4439 @kyleniemeyer
@fabian-s It looks like an ISBN (or multiple) may be incorrect and causing the error above. Can you remove those for the journal articles where they are present? That metadata is not necessary for journal articles anyways
