### 2022-06-28 b55afcc
Increment version number to 1.2.3
