+++ NEWS.md
@@ -1,3 +1,5 @@
+# musclesyneRgies 1.2.3
+
 # musclesyneRgies 1.2.2
 ### How to install
 ```
