### 2022-06-28 60c559c
Corrected JOSS paper bibliography
