+++ paper/paper.bib
@@ -84,7 +84,6 @@
 @article{Tresch1999,
    author = {Matthew C. Tresch and Philippe Saltiel and Emilio Bizzi},
    doi = {10.1038/5721},
-   isbn = {1097-6256},
    issn = {1097-6256},
    issue = {2},
    journal = {Nature Neuroscience},
@@ -98,7 +97,6 @@
 @article{Bruton2018,
    author = {Michaela Bruton and Nicholas O'Dwyer},
    doi = {10.1152/jn.00052.2018},
-   isbn = {00052-2018},
    issn = {0022-3077},
    issue = {6},
    journal = {Journal of Neurophysiology},
@@ -191,7 +189,6 @@
 @article{Martino2014,
    author = {Giovanni Martino and Yury P. Ivanenko and M. Serrao and A. Ranavolo and Andrea D'Avella and F. Draicchio and C. Conte and C. Casali and Francesco Lacquaniti},
    doi = {10.1152/jn.00275.2014},
-   isbn = {1522-1598 (Electronic)\r0022-3077 (Linking)},
    issn = {0022-3077},
    issue = {11},
    journal = {Journal of Neurophysiology},
@@ -245,7 +242,6 @@
 @article{Rosenstein1993,
    author = {Michael T. Rosenstein and James J. Collins and Carlo J. De Luca},
    doi = {10.1016/0167-2789(93)90009-P},
-   isbn = {0167-2789(92)00033-6},
    issn = {01672789},
    issue = {1-2},
    journal = {Physica D},
