### 2022-06-28 #4439 @alesantuz
@kyleniemeyer and @fabian-s I removed the ISBNs in the .bib file.
The new version 1.2.3 is now archived on GitHub [here](https://github.com/alesantuz/musclesyneRgies/releases/tag/v1.2.3) and on Zenodo [here](https://doi.org/10.5281/zenodo.6767313) with DOI 10.5281/zenodo.6673110.
Thank you @kyleniemeyer for looking into this!
