### 2022-06-28 #4439 @fabian-s
"missing" DOI suggested above is wrong, the paper is from 1967 and does not have a DOI.
